+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Tip
+
+* Go float64 can no longer be encoded to a PostgreSQL float4
+* Add ConnPool.Reset method
+* []byte skips encoding/decoding
+* Rows.Scan errors now include which argument caused error
+
 # 2.7.1 (October 26, 2015)
 
 * Disable SSL renegotiation